+++ formatters/date.js
@@ -29,7 +29,97 @@ var dayjs = require('dayjs');
  */
 function formatD (d, patternOut, patternIn) {
   if (d !== null && typeof d !== 'undefined') {
-    return fromMomentToDayJS(d, patternIn).tz(this.timezone).locale(this.lang).format(patternOut);
+    return parse(d, patternIn).tz(this.timezone).locale(this.lang).format(patternOut);
+  }
+  return d;
+}
+
+/**
+ * Adds days, hour, year, ... to a date
+ *
+ * @exampleContext {"lang":"fr", "timezone":"Europe/Paris"}
+ * @example ["2017-05-10T15:57:23.769561+03:00", "3", "day"]
+ * @example ["2017-05-10 15:57:23.769561+03:00", "3", "month"]
+ * @example ["20160131", "3", "day"]
+ * @example ["20160131", "3", "month"]
+ * @example ["31-2016-01", "3", "month", "DD-YYYY-MM"]
+ *
+ * @param      {String|Number}  d   input date
+ * @param      {Number}  amount     The amount
+ * @param      {String}  unit       The unit
+ * @param      {String}  patternIn  [optional] input format, ISO8601 by default
+ * @return     {Date}               return a date, which can be formatted with formatD, or manipulated with other formatters
+ */
+function addD (d, amount, unit, patternIn) {
+  if (d !== null && typeof d !== 'undefined') {
+    return parse(d, patternIn).add(parseInt(amount, 10), unit || 'day');
+  }
+  return d;
+}
+
+/**
+ * Subtract days, hour, year, ... to a date
+ *
+ * @exampleContext {"lang":"fr", "timezone":"Europe/Paris"}
+ * @example ["2017-05-10T15:57:23.769561+03:00", "3", "day"]
+ * @example ["2017-05-10 15:57:23.769561+03:00", "3", "month"]
+ * @example ["20160131", "3", "day"]
+ * @example ["20160131", "3", "month"]
+ * @example ["31-2016-01", "3", "month", "DD-YYYY-MM"]
+ *
+ * @param      {String|Number}  d   input date
+ * @param      {Number}  amount     The amount
+ * @param      {String}  unit       The unit
+ * @param      {String}  patternIn  [optional] input format, ISO8601 by default
+ * @return     {Date}               return a date, which can be formatted with formatD, or manipulated with other formatters
+ */
+function subD (d, amount, unit, patternIn) {
+  if (d !== null && typeof d !== 'undefined') {
+    return parse(d, patternIn).subtract(parseInt(amount, 10), unit || 'day');
+  }
+  return d;
+}
+
+/**
+ * Set a date to the start of a unit of time.
+ *
+ * @exampleContext {"lang":"fr", "timezone":"Europe/Paris"}
+ * @example ["2017-05-10T15:57:23.769561+03:00", "day"]
+ * @example ["2017-05-10 15:57:23.769561+03:00", "month"]
+ * @example ["20160131", "day"]
+ * @example ["20160131", "month"]
+ * @example ["31-2016-01", "month", "DD-YYYY-MM"]
+ *
+ * @param      {String|Number}  d   input date
+ * @param      {String}  unit       The unit
+ * @param      {String}  patternIn  [optional] input format, ISO8601 by default
+ * @return     {Date}               return a date, which can be formatted with formatD, or manipulated with other formatters
+ */
+function startOfD (d, unit, patternIn) {
+  if (d !== null && typeof d !== 'undefined') {
+    return parse(d, patternIn).startOf( unit || 'year');
+  }
+  return d;
+}
+
+/**
+ * Set a date to the end of a unit of time.
+ *
+ * @exampleContext {"lang":"fr", "timezone":"Europe/Paris"}
+ * @example ["2017-05-10T15:57:23.769561+03:00", "day"]
+ * @example ["2017-05-10 15:57:23.769561+03:00", "month"]
+ * @example ["20160131", "day"]
+ * @example ["20160131", "month"]
+ * @example ["31-2016-01", "month", "DD-YYYY-MM"]
+ *
+ * @param      {String|Number}  d   input date
+ * @param      {String}  unit       The unit
+ * @param      {String}  patternIn  [optional] input format, ISO8601 by default
+ * @return     {Date}               return a date, which can be formatted with formatD, or manipulated with other formatters
+ */
+function endOfD (d, unit, patternIn) {
+  if (d !== null && typeof d !== 'undefined') {
+    return parse(d, patternIn).endOf(unit || 'year');
   }
   return d;
 }
@@ -60,14 +150,20 @@ function convDate (d, patternIn, patternOut) {
   return formatD.call(this, d, patternOut, patternIn);
 }
 
+
 /**
  * Convert old MomentJS format to DayJS format
  *
+ * @private
  * @param      {string}  d          not undefined/null date
  * @param      {string}  patternIn  The pattern
  * @return     {Object}             dayjs
  */
-function fromMomentToDayJS (d, patternIn) {
+function parse (d, patternIn) {
+  // if the date is already parsed
+  if (typeof(d) === 'object' && d.isValid) {
+    return d;
+  }
   if (!patternIn) {
     return dayjs(d + '');
   }
@@ -77,12 +173,16 @@ function fromMomentToDayJS (d, patternIn) {
   if (patternIn === 'x') {
     return dayjs(d);
   }
-  return dayjs(d + '', patternIn);
+  return dayjs(d, patternIn);
 }
 
 
 
 module.exports = {
-  formatD  : formatD,
-  convDate : convDate
+  formatD,
+  convDate,
+  addD,
+  subD,
+  startOfD,
+  endOfD
 };