+++ CHANGELOG.md
@@ -1,20 +1,46 @@
 
 ### v2.2.0
   - ⚡️ **Manage timezone + new date formatters + switch from MomentJS to DayJS**
+    - If not defined by you in `options.complement`, `{c.now}` returns the current date in UTC.
     - [BREAKING CHANGE]: remove old date formatter which were not documented: `format`, `parse`, `addDays` and `convert`.
-      You should use `formatD` instead and new formatters below. It was very old formatters, the chance you use them is low because you had to 
-      look in the code to know their existance.
-    - [BREAKING CHANGE]: We try to stay as close as possible as the previous parsing algorithm. But we have noticed some changement:
-      - `1997-12-17 07:37:16-08:00` ✅ accepted, same as previous version
-      - `1997-12-17 07:37:16Z`      ✅ accepted, same as previous version
-      - `1997-12-17 07:37:16-08`    ❌ not accepted anymore without specifying an input pattern
-    - Manage timezone: https://en.wikipedia.org/wiki/List_of_tz_database_time_zones
-    - TODO show a complete example 
-    ```js
-      options = {
-        timezone: 'Europe/Paris'
-      }
-    ```
+      You should use `formatD` instead and new formatters below. They were very old formatters, the chance you use them is low because you had to 
+      look into the source code to know their existance.
+    - New formatters:
+      - `addD(amount, unit [, patternIn])`     : add days, month to a date. `formatD` can be used after without specifying  patternIn 
+      - `subD(amount, unit [, patternIn])`     : subtract days, month to a date. `formatD` can be used after without specifying  patternIn
+      - `startOfD(amount, unit [, patternIn])` : Set a date to the start of a unit of time. `formatD` can be used after without specifying  patternIn
+      - `endOfD(amount, unit [, patternIn])`   : Set a date to the end of a unit of time. `formatD` can be used after without specifying  patternIn
+    - [BREAKING CHANGE]: We try to stay as close as possible as the previous parsing algorithm.
+      But `1997-12-17 07:37:16-08` is not accepted anymore without specifying an input pattern or writing `1997-12-17 07:37:16-08:00`
+    - Accept a new options to set the `timezone` of `formatD`. Examples:
+
+      *Data*
+      ```js
+        {
+          date     : '2020-11-28T21:54:00.000Z',  // ISO 8601 UTC
+          dateWTZ  : '2020-11-28T21:54:00',       // without timezone
+          dateTZ   : '2020-11-28T21:54:00-04:00', // with America/New_York timezone offset
+          dateUnix : 1606600440                   // UNIX timestamp in seconds UTC of 2020-11-28T21:54:00.000Z
+        }
+      ```
+
+      *Template => Result*
+
+      if `options.timezone = 'Europe/Paris'` (default)
+        - `{d.date:formatD(LLLL)}`        => `Saturday, November 28, 2020 10:54 PM`
+        - `{d.dateWTZ:formatD(LLLL)}`     => `Saturday, November 28, 2020 9:54 PM`
+        - `{d.dateTZ:formatD(LLLL)}`      => `Sunday  , November 29, 2020 2:54 AM`
+        - `{d.dateUnix:formatD(LLLL, X)}` => `Saturday, November 28, 2020 10:54 PM`
+
+      if `options.timezone = 'America/New_York'` 
+        - `{d.date:formatD(LLLL)}`        => `Saturday, November 28, 2020 4:54 PM`
+        - `{d.dateWTZ:formatD(LLLL)}`     => `Saturday, November 28, 2020 3:54 PM`
+        - `{d.dateTZ:formatD(LLLL)}`      => `Saturday, November 28, 2020 8:54 PM`
+        - `{d.dateUnix:formatD(LLLL, X)}` => `Saturday, November 28, 2020 4:54 PM`
+
+      List of timezone:  https://en.wikipedia.org/wiki/List_of_tz_database_time_zones 
+
+
   - Fix: if a path does not exist inside a formatter argument, it returns an empty string instead of the error "[[C_ERROR]] attribute_name not defined".
     It fixes some weird behaviour with ifEM formatters
   - Accepts to convert the first page of docx or odt templates into a JPEG file with `converTo : 'jpg'`